requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
requests>=2.31.0
tqdm>=4.65.0
biopython>=1.81
pyseer>=1.3.11
matplotlib>=3.7.0
seaborn>=0.12.0
adjustText>=0.8
pyyaml>=6.0
loguru>=0.7.0
joblib>=1.3.0
